separating client connection and adding logger
